+++ src/index.ts
@@ -1,14 +1,19 @@
 export * from "./openscad/OpenScad.js";
-export * from "./openscad/IOpenScadOptions.js";
-export * from "./openscad/OpenScadOutput.js";
-export * from "./openscad/ParameterSet.js";
+export * from "./openscad/ParameterSetLoader.js";
+export * from "./openscad/Summary.js";
 
 export * from "./schemas/openscadOptionsSchema.js";
 export * from "./schemas/openscadParameterDefinitionSchema.js";
 export * from "./schemas/openscadParameterKvSchema.js";
 
+export * from "./types/IOpenScad.js";
+export * from "./types/IOpenScadOptions.js";
+export * from "./types/OpenScadSummary.js";
+export * from "./types/ParameterDefinition.js";
+export * from "./types/ParameterSet.js";
+export * from "./types/RecursivePartial.js";
+
 export * from "./util/execBash.js";
 export * from "./util/AnimationGeneration.js";
 export * from "./util/MosaicGeneration.js";
-export * from "./util/RecursivePartial.js";
 export * from "./util/validation.js";