+++ src/types/ParameterSet.ts
@@ -1,5 +1,3 @@
-import fs from "node:fs";
-
 export class ParameterSet {
   public parameterSets: Record<string, Record<string, string>> = {};
   public fileFormatVersion: "1" = "1" as const;
@@ -8,20 +6,6 @@ export class ParameterSet {
     if (p) {
       Object.assign(this, p);
     }
-  }
-
-  public static createFromFile(filePath: string): ParameterSet {
-    return JSON.parse(fs.readFileSync(filePath, "utf8")) as ParameterSet;
-  }
-
-  public static writeToFile(filePath: string) {
-    fs.writeFileSync(filePath, JSON.stringify(this, null, 2), "utf8");
-  }
-
-  public static toParameterSet(paramKV: ParameterKV[], name: string = "model"): ParameterSet {
-    const parameterSet = new ParameterSet();
-    parameterSet.add(name, paramKV);
-    return parameterSet;
   }
 
   public add(name: string, paramKV: ParameterKV[]) {